+++ src/App.jsx
@@ -1,15 +1,29 @@
-import { MoviesGrid } from './MoviesGrid';
-import styles from './App.module.css';
+import styles from "./App.module.css";
+import {
+  BrowserRouter as Router,
+  Routes,
+  Route,
+  Link,
+  Navigate,
+} from "react-router-dom";
+import { MovieDetails } from "./pages/MovieDetails";
+import { LandingPage } from "./pages/LandingPage";
 
 export function App() {
-	return(
-		<div>
-			<header>
-				<h1 className={styles.title}>Movies</h1>
-			</header>
-			<main>
-				<MoviesGrid />
-			</main>
-		</div>
-	);	
+  return (
+    <Router>
+      <header>
+        <Link to="/">
+          <h1 className={styles.title}>Movies</h1>
+        </Link>
+      </header>
+      <main>
+        <Routes>
+          <Route path="/movies/:movieId" element={<MovieDetails />} />
+          <Route path="/" element={<LandingPage />} />
+          <Route path="*" element={<Navigate replace to="/" />} />
+        </Routes>
+      </main>
+    </Router>
+  );
 }
