+++ src/pages/LandingPage.jsx
@@ -1,0 +1,7 @@
+import { MoviesGrid } from '../components/MoviesGrid';
+
+export function LandingPage() {
+	return(
+		<MoviesGrid />
+	);
+}
