+++ src/pages/MovieDetails.jsx
@@ -1,0 +1,7 @@
+export function MovieDetails() {
+	return(
+		<div>
+			Details
+		</div>
+	);
+}
